Rename satpy.data_download to satpy.aux_download
