+++ satpy/readers/yaml_reader.py
@@ -42,7 +42,7 @@ from satpy.resample import get_area_def
 from satpy.utils import recursive_dict_update
 from satpy.dataset import DataQuery, DataID, get_key
 from satpy.dataset.dataid import get_keys_from_config, default_id_keys_config, default_co_keys_config
-from satpy.data_download import DataDownloadMixin
+from satpy.aux_download import DataDownloadMixin
 from satpy import DatasetDict
 from satpy.resample import add_crs_xy_coords
 from trollsift.parser import globify, parse
